requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-groq
google-genai
faiss-cpu
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
pydantic
python-dotenv
numpy
hypothesis
pytest
